use flexbox, split css
