+++ packages/jsenv-plugin-toolbar/src/client/toolbar.html
@@ -382,7 +382,7 @@
       </div>
 
       <div id="overflow-menu">
-        <div style="width: 100%"></div>
+        <div style="flex: 1"></div>
       </div>
 
       <div id="overflow-menu-button">
