+++ packages/jsenv-plugin-toolbar/src/client/toolbar_core.js
@@ -1,12 +1,10 @@
 import { updateToolbarState } from "./toolbar_state.js"
-import { addExternalCommandCallback } from "./communication/parent_window_communication.js"
+import { addExternalCommandCallback } from "./core/parent_window_communication.js"
 import { renderToolbar, showToolbar, hideToolbar } from "./ui/toolbar_ui.js"
 
 // const { currentScript } = document
 addExternalCommandCallback("renderToolbar", ({ logs }) => {
-  renderToolbar({
-    logs,
-  })
+  renderToolbar({ logs })
 })
 addExternalCommandCallback("showToolbar", () => {
   showToolbar()
