+++ packages/jsenv-plugin-toolbar/src/client/ui/layout.css
@@ -1,0 +1,26 @@
+#toolbar-overlay {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  filter: url(#better-blur);
+  background-color: rgba(0, 0, 0, 0.2);
+}
+
+#toolbar {
+  position: fixed;
+  bottom: 0;
+  height: 40px;
+  width: 100%;
+  background: #404040;
+}
+
+#toolbar-wrapper {
+  display: flex;
+  justify-content: center;
+}
+
+/* #toolbar-wrapper > * {
+  vertical-align: middle;
+} */
