+++ packages/jsenv-plugin-toolbar/src/client/ui/toolbar.css
@@ -1,69 +1,13 @@
+@import "./globals.css";
+@import "./layout.css";
+@import "./toolbar_trigger.css";
 @import "../execution/execution.css";
-@import "../focus/focus.css";
+@import "./focus.css";
 @import "../eventsource/eventsource.css";
 @import "../responsive/overflow_menu.css";
 @import "../settings/settings.css";
 @import "../theme/light_theme.css";
 @import "../tooltip/tooltip.css";
-
-html[data-animation-disabled] * {
-  transition: none !important;
-  animation-duration: 0s !important;
-  animation-delay: 0s !important;
-}
-
-[data-when] {
-  display: none !important;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-}
-
-html {
-  font-family: arial, sans;
-}
-
-/* global utils */
-[data-force-hide] {
-  display: none !important;
-}
-
-#toolbar-overlay {
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  filter: url(#better-blur);
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
-#toolbar {
-  position: fixed;
-  bottom: 0;
-  height: 40px;
-  width: 100%;
-  background: #404040;
-}
-
-#toolbar-wrapper {
-  display: table;
-  width: 100%;
-  height: 100%;
-}
-
-#toolbar-wrapper > * {
-  display: table-cell;
-  vertical-align: middle;
-}
 
 #toolbar button {
   background: none;
@@ -118,50 +62,6 @@ html {
 /* close icon */
 #crossIconSvg {
   width: 18px;
-}
-
-/* re-open toolbar-trigger */
-#toolbar-trigger {
-  display: block;
-  overflow: hidden;
-  position: absolute;
-  bottom: -32px;
-  right: 20px;
-  height: 40px;
-  width: 40px;
-  padding: 0;
-  border-radius: 5px 5px 0 0;
-  border: 1px solid rgba(0, 0, 0, 0.33);
-  border-bottom: none;
-  box-shadow: 0px 0px 6px 2px rgba(0, 0, 0, 0.46);
-  background: transparent;
-}
-
-#toolbar-trigger[data-animate] {
-  transition: 600ms;
-}
-
-#toolbar-trigger #jsenvLogoSvg {
-  width: 35px;
-  height: 35px;
-  opacity: 0;
-  transition: 600ms;
-}
-
-html[data-toolbar-visible] #toolbar-trigger {
-  display: none;
-}
-
-#toolbar-trigger:hover {
-  cursor: pointer;
-}
-
-#toolbar-trigger[data-expanded] {
-  bottom: 0;
-}
-
-#toolbar-trigger[data-expanded] #jsenvLogoSvg {
-  opacity: 1;
 }
 
 /* #jsenvLogo {
