+++ packages/jsenv-plugin-toolbar/src/client/ui/toolbar_ui.js
@@ -54,8 +54,9 @@ export const renderToolbar = async () => {
   // if user click enter or space quickly while closing toolbar
   // it will cancel the closing
   // that's why I used toggleToolbar and not hideToolbar
-  document.querySelector("#button-close-toolbar").onclick = () =>
+  document.querySelector("#button-close-toolbar").onclick = () => {
     toogleToolbar()
+  }
 }
 
 const toogleToolbar = () => {
