+++ src/plugins/explorer/client/explorer.html
@@ -175,6 +175,7 @@
         padding-right: 25px;
         padding-top: 25px;
         background-color: #204143;
+        position: relative;
       }
 
       #explorables_header_bottom_spacing {
@@ -325,7 +326,7 @@
         <section id="explorables">
           <div id="explorables-header">
             <div id="explorables_header_and_menu">
-              <h2 style="white-space: nowrap">
+              <h2>
                 <svg
                   id="fileIconSvgConfig"
                   viewBox="0 0 24 24"
